+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `phone()` now pre-processes values that are considered "not a number" to avoid running them through the Java parser. This massively speeds up phone number parsing for `""`, `NA` and similar values, and fixes an issue where Java would freeze after seeing a large amount of invalid phone number values (#27).
 
+* It is now possible to create empty `phone()` vectors (#28).
+
 * `show_progress` is now a function argument for `phone()` and `get_timezone()`. Previously progress bar visibility could only be set package wide using the `dialr.show_progress` option. Now it can be set locally using the `show_progress` function parameter (#25).
 
 # dialr 0.4.0